Removing "Windows is not officially tested"
According to the dev, this is no longer the case.  We did test the integration with Notary for Windows images.  Let me know if you need more information around this.
 
"Windows is not officially tested, but if you are a developer and Windows user,
we would appreciate any insight you can provide regarding issues."