Docker for Mac: add page on disk space
The disk space usage of Docker on Mac has been difficult to manage
and has confused lots of people, see

https://github.com/docker/for-mac/issues/371

This patch adds a top-level section to the Docker for Mac docs
alongside networking which explains

- where the space has gone
- how to reclaim some of it

In future I'd like to add more about why some systems use `Docker.raw`
and some `Docker.qcow2` but we're in a transitional state at the moment.
It'll be simpler to explain after the next stable version has been
released.

Signed-off-by: David Scott <dave.scott@docker.com>