Insert shell env var into env var priority order
As mentioned in lines 115 through 117, "Values in the shell take precedence over those specified in the `.env` file. If you set `TAG` to a different value in your shell, the substitution in `image` uses that instead:".

This context should also be represented in the order of priority in which variables are processed.