Correcting command to follow text
The text says that you want to stop the container and that it will be automatically removed due to being started with the --rm option.  But the command that is given container remove command.  Even if the container was not started with the --rm option, that command will not work because you must stop the container before removing it.