linked ucp user bundle and clarified --ucp-node
Linked the UCP user bundle to instructions on how to download and use a UCP user bundle. Added a note to clarify that --ucp-node is the node you are trying to install DTR on. See a lot of people setting this to one of their UCP manager nodes by accident or ignorance.