Refer Windows Subsystem for Linux for Windows 10
Hi, 
this is a great guide and I am just following along. Got to this point and found old references to getting bash on windows. So, adding a MSDN reference to install the Windows Subsystem for Linux - https://docs.microsoft.com/en-us/windows/wsl/install-win10. Since this is commonly available, updating the guide with it. there should be no need to install extra software and is a fantastic utility that helps developers using Windows 10 learning docker.. 
Some developers may have enterprise IT constraints and hence may still require the gitbash or wget software and hence retained those references as well.. 
thanks rasane