clear up confusion between compose file versions
Really this type of info could be placed elsewhere and referenced in this note, so maybe suggest where it would live better. See comments https://github.com/moby/moby/issues/31101#issuecomment-428472115 and others for how we could better communicate the "best fit" file versions depending on someones use case. This is a top 3 issue of anyone I meet trying to use compose files with or without swarm. They are all confused thinking 2.x branch is not supported or maintained now that 3.x exists.