Added `demo_` prefix to interlock label
Customer feedback received: 

```
It leads users to expect that if they define a network in their compose file that they can use that network name in the Interlock network label which is not the case.  The example "works" because the service is only attached to a single network, but in actuality the Interlock network label would need to be prefixed with the stack name in order for the example to work correctly.
```
The concern is valid; so I have added `demo_` as a prefix to the `com.docker.lb.network` label.