Dockerfile: split stages for "dev" and "deploy"
The current Dockerfile defaults to include all archived documentation. While this
is the right thing to do when _deploying_ the documentation, it is far from optimal
when working on changes in the documentation, and building the documentation locally
(in which case, the archived documentation is likely not interesting).

This patch changes the Dockerfile to only build the current version of the documentation
by default.

> **Note** when building with the "classic" builder (so not using BuildKit), the archived
> documentation will still be built, because the classic builder is not able to skip
> unused build-stages.

This change might require changes in how the documentation is deployed, because in that
case, the documentation _including_ archives should be built.

Alternatively, the default could be changed to include the archives, but people wanting
to build docs without archives could set `--target=dev` when running `docker build`.

Before this change a full rebuild of the docs, including archives, took ~247 seconds
(already optimised if BuildKit is used, which fetches the archived docs in parallel):

```
docker build -t docs --target=deploy .

[+] Building 247.1s (30/30) FINISHED
 => FROM docker.io/docs/docker.github.io:v18.03                                                                                                                                                                                     151.4s
 => FROM docker.io/docs/docker.github.io:v17.03                                                                                                                                                                                      75.0s
 => FROM docker.io/docs/docker.github.io:v17.09                                                                                                                                                                                      96.5s
 => FROM docker.io/docs/docker.github.io:v17.06                                                                                                                                                                                     134.9s
 => FROM docker.io/docs/docker.github.io:v17.12                                                                                                                                                                                     105.2s
...
 => [archives 1/5] COPY --from=docs/docker.github.io:v17.03 /usr/share/nginx/html /usr/share/nginx/html                                                                                                                               1.7s
 => [archives 2/5] COPY --from=docs/docker.github.io:v17.06 /usr/share/nginx/html /usr/share/nginx/html                                                                                                                               2.8s
 => [archives 3/5] COPY --from=docs/docker.github.io:v17.09 /usr/share/nginx/html /usr/share/nginx/html                                                                                                                               3.8s
 => [archives 4/5] COPY --from=docs/docker.github.io:v17.12 /usr/share/nginx/html /usr/share/nginx/html                                                                                                                               4.6s
 => [archives 5/5] COPY --from=docs/docker.github.io:v18.03 /usr/share/nginx/html /usr/share/nginx/html                                                                                                                               2.7s
```

The size of the built image is ~1.6 GB

```
REPOSITORY              TAG                      IMAGE ID            CREATED             SIZE
docs                    latest                   8fd1a9d31da7        4 minutes ago       1.61GB
```

With this change, the archived docs are not fetched, and a full rebuild of the docs is reduced to ~200 seconds

```
[+] Building 203.1s (20/20) FINISHED
```

Given that the archives are not included by default, the size of the image also reduced considerably,
and is now about 340MB

```
REPOSITORY              TAG                      IMAGE ID            CREATED             SIZE
docs                    latest                   bb7b1b9af649        7 seconds ago       339MB
```

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>