Adding Release Notes commentary (#1161)
* Adding Release Notes commentary

Per HUB-1827, we discovered a Hub feature which is *very valuable* to partners & customers, but was not previously documented. Given that we only discovered this because _one partner_ was made aware by _one partner manager_, it is a great example of tribal knowledge we can disseminate via proactive documentation. 

Adding to the FAQ's now so it will not be lost / forgotten. In a more comprehensive revision, this could easily be situated in a dedicated "Publisher Portal Usage Docs" section.

* Minor edit