Revising antiquated product references, deleting incorrect statements (#1152)
https://github.com/docker/docs-private/issues/1151
* Docker Store references
* Incorrect DCI references
* Docker Enterprise Edition references
* Removing "Private repo" requirement which is _not actually true and has mislead / frustrated many partners_
* Removing SLA - This has never been enforced, to the extent that we don't even have Certification requirements around resolving the CVE's anymore!
* Deleted DCI reference, because it is wildly inaccurate :)
* Program Guide URL
* Listing fee is a lie.. we've kept that up for 3+ years, time to let it go!
* Removing old versioning which is irrelevant 2 years later..

Just a note... much of this look extremely redundant with the Publisher FAQ's. Also, there's no link to this page at all within the doc's tree. There's only a lone link to it at the end of another doc's section.. not sure if this page was intended to be shut down? If not, we ought to get it pulled back into the tree again.