Update TOC and change "overview" to "index"
This changes the index page URL for cluster to `/cluster/`
instead of `/cluster/overview/`, which makes URL navigation
more fitting the expectations.

Note that there's still issues to resolve; because we're adding links
to the *same* "reference" pages and "environment variables" in multiple
locations in the navigation tree, the navigation marks all occurrences
as "active". Which means that browsing the Docker CLI reference, and
going to the "cluster" section, will make the navigation jump to the
"Deploy Docker Enterprise" section.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>