Share redoc JavaScript between API versions
This is similar to the changes made in  docker/docker.github.io#5607
(32958e7c609efdc7507c5eb215553447c2024d2a), but which appear
to didn't make their way into newer versions that were added.

The redoc JavaScript is a big file, and there's no need
to keep a separate copy for each version of the API.

This patch removes the copy per-version and moves the script
to a central location.

Size of the API directory:

before: 5.961.110 bytes (6 MB on disk)
after:     15.570 bytes (74 KB on disk)

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>