Removed extraneous azure.json entries (#9271)
Currently we specify quite a few extra lines in the `azure.json` file that are unneeded due to the default values. 

I removed `cloudProviderBackoff` and its extraneous related values, and `useManagedIdentityExtension` since the default value is the same. 

Removing these entries tidies up and simplifies the sample `azure.json` for end users.