Added example command for `docker context import` (#9423)
Added an actual `docker context import` command since the use of this command in conjunction with client bundles was not present in this or the `docker context` documentation. Also increased the emphasis of using `docker context import` vs. prior methods by pulling the new feature out of the note. This information is important and was not obvious enough in the note.