logging/gelf: recommend to disable compression
It was discovered that compression in gelf logging driver
is highly inefficient in terms of CPU. IOW, lots of CPU
is used when compression is enabled.

The natural solution would be to change the default
to `none`, i.e. no compression, but that would break
some setups (in particular, using Logstash < 7.4).

The only option left is to warn the user about the
consequences of leaving the compression enabled.

For more background details, please see
https://github.com/moby/moby/pull/40101

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>