Clarify healthcheck CMD-SHELL command
With the current description, it is not clear that for example setting this array as test
test: ["CMD-SHELL", "curl -s -X GET http://127.0.0.1:5000/actuator/health", "|", "grep 'UP'"]

would result in executing the first (curl) command only to evaluate the status. (more details https://stackoverflow.com/questions/59269073/docker-swarm-stack-file-healthcheck-command-does-not-work-as-expected/59390396#59390396)
It is not clear because it is said that `test must be either a string or a list. If it’s a list, the first item must be either NONE, CMD or CMD-SHELL` - so it does not make sense to use CMD-SHELL with a list.