Adding untagged parameter description
Based on the pull request https://github.com/docker/distribution/pull/2302 there is a parameter for cleaning up untagged manifests available. In the pull request it is called "-m" / "--delete-manifests", but in the current version of registry and it's garbage-collection, it's called "--delete-untagged" (checked with "registry garbage-collect -h" inside the latest registry-image based container). Anyway, it was missing from the documentation so I thought it would be useful to add here.