Unify API documentation styles
Docker Engine, DTR, and UCP all use a swagger (OpenAPI) file to document
the API, but different javascript libraries and styles were used.

This patch unifies these API docs to use the same library, which reduces
maintenance, reduces the amount of data to load in the browser (the
javascript and css can be cached), and allows updating the style in
a single place (keeping a consisten style).

To test this change;

1. build the docs

    docker build -t docs .

2. run the docs

    docker run -it --rm -p 4000:4000 docs

3. visit the API docs in your browser

    - DTR: http://localhost:4000/reference/dtr/2.7/api/#operation/DeleteNamespace
    - UCP: http://localhost:4000/reference/ucp/3.2/api/#tag/Container

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>