Remove highlight.js
Code is already highlighted through "rouge", so enabling highlight.js
only resulted in code being parsed/highlighted _twice_.

Highlight.js was only included on pages that explicitly enabled it,
which was not used anywhere, so removing it should not have an effect.

This patch removes highlight.js and updates the "github" stylesheet to
be compatible with rouge.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>