Fix broken link in mac FAQs (#10324)
* Fix broken link in mac FAQs

The title of the linked section has changed to "Add TLS Certificates". The current link goes to the top of the Getting Started page and is confusing because you have to scroll past the Kubernetes section to get to the section you actually need.

* Fix broken link to mac FAQs

I missed removing `/index` in the first commit. This is the correct link. Thanks to thaJeztah for the correction.

Co-Authored-By: Sebastiaan van Stijn <thaJeztah@users.noreply.github.com>

Co-authored-by: Sebastiaan van Stijn <thaJeztah@users.noreply.github.com>