Update engine cli reference to fix some missing descriptions
Regenerated the yaml files after fixing a bug in the generator
script that caused some extended descriptions and examples
to not be included.

Also fixes the generated YAML to use the "long form" format,
instead of the compact format (where newlines were encoded
as `\n`). This makes the YAML more "human readable", and makes
reviewing updates easier.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>