Changed all 'apt-get" to "apt"
As of, you state in there, this is for: Debian 10 or Raspbian 9, it is recommended for those to use "apt" not "apt-get".
apt is 100% compatible with apt-get even in case of repos and other such as: add-apt-key, apt-cache etc; I've tested that installing Docker Client and Docker Engine on my Linux Mint 19.3 host. - replacing apt-get to apt works very well.