Fix curl examples.md
When running with curl without the --output parameter it shows an error and it doesn't show the actual container logs:

Warning: Binary output can mess up your terminal. Use "--output -" to tell 
Warning: curl to output it to your terminal anyway, or consider "--output 
Warning: <FILE>" to save to a file.