Removed wrong information about container DNS
In the docs it says a container will inherit the host /etc/hosts file along with /etc/resolv.conf. While the latter part is correct, /etc/hosts will not actually be inherited, as reported https://github.com/moby/moby/issues/39311. 
An additional line of text has been added to emphasize the fact that /etc/hosts are not inherited