Remove volume configuration
Since this setting is not required for Quickstart.
If it requires database state,
it should not remove container but stop it.
When restart container, data is kept by the benefit of docker volume.
Or, almost programming language can define test data by code
and reproduce it by using fixture library like factory_bot.