Specify both the major and minor numbers
Specify both the major and minor numbers Compose file version.
If no minor version is given, 0 is used by default
and not the latest minor version.
This behavior will confuse beginners
and it has possibility for a beginner
to mistake the wrong way for the right way.

cf. https://docs.docker.com/compose/compose-file/compose-versioning/#v2-and-v3-declaration