_includes/head.html: generate page title from H1 if needed
This is a very hacky way to extract the page title from pages that do not have
front-matter yaml, but have a H1 header. We need to take (id-) attributes into
account, so some hacking is needed.

Note that there's also a Jekyll plugin that features similar functionality, but
it requires additional dependencies, and we only have a few pages that need
this, so for now using this hack.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>