Remove glossary left-overs
Looks like removing the front-matter in f17ebae568a15033c914d2024f5e939dee89a4a8
caused the output to break, resulting in a JavaScript error.

Looking at where this file was used, it turned out that it was loaded, but
never used anywhere.

This commit removes the remaining parts of the glossary search functionality,
which was not used.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>