Remove highlight.js script as it was not enabled
Code is already highlighted through "rouge", so enabling highlight.js
only resulted in code being parsed/highlighted _twice_.

Highlight.js was only included on pages that explicitly enabled it,
which was not used anywhere, so removing it should not have an effect.

This patch removes highlight.js. There are some stylesheets that
can be removed and/or merged after this, but leaving that separate.

The github.css stylesheet is already included in the style.css
stylesheet (from the _scss directory), so was redundant.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>