engine/examples: remove sshd example
The ssh example was really bad; it even leaks the root password as a plain text in the image :(

This commit removes the sshd example and discourage people to dockerize sshd.

The `running_ssh_service.md` page itself is not removed because there
are hyperlinks on the Internet to that page.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>