Update ubuntu.md
### Why this PR?

I got confused. Is setting up the docker engine part of the instructions for installing via a repository? It's not easy to distinguish between the level 3 and level 4 markup headings. I thought I would number them explicitly to avoid confusion. A small change, but hopefully it is helpful. if not, pls feel free to close. chrs.

Ben