Update Debian install documentation
"the 64-bit version of one of these Debian or Raspbian versions" might be confusing, since there is no 64-bit Raspbian, which is naturally an armv6hf (32-bit) only repository. The 64-bit Raspberry Pi OS uses the regular arm64 Debian repository. Remove the 64-bit mention from this sentence, as the supported architectures are listed below. Add "Raspbian Buster" explicitly to the supported OS list.

Do not use the deprecated "apt-key" command to add the repository key. Use "gpg --dearmor" instead to install the key as separate file into /etc/apt/trusted.gpg.d/. This fixes issue #11851.
The deprecation warning can be seen e.g. in the Debian man pages: https://manpages.debian.org/testing/apt/apt-key.8.en.html

Do not use the "add-apt-repository" command to add the repository list. It does not work on Raspbian, due to a missing template file: https://github.com/RPi-Distro/repo/issues/160
Moreover, the passed command line arguments are the exact content that needs to be added to the list file, hence can be used with a simple "echo". On Raspbian, the URL path needs to be adjusted. Since "lsb_release" cannot be used to print the distribution ID in lower case, instructions for Debian and Raspbian are separated into two tabs. The three tabs for the three supported package architectures have been merged via "$(dpkg --print-architecture)". A note has been added to replace this command substitution, if a foreign architecture is wanted. It may be removed completely to let DPkg choose the native architecture automatically, or the next best added foreign architecture, if the native one is not provided by the repository. With this change, the method can be used on Raspbian, hence the information about which method is supported by which distribution, has been removed.

Do not expect the "lsb_release" command to be present, add the "lsb-release" package to the dependency list instead. "software-properties-common" is not required anymore as "add-apt-repository" is not used. To assure that the "gpg" command is available and the key can be handled by the system, add the "gnupg" package as well.

The "apt-cache madison docker-ce" example output has been updated to show current packages available on Debian Buster.

To support Debian and Raspbian, the MkDocs "download-url-base" variable has been truncated and "/debian" resp. "/raspbian" needs to be added literally to when calling it.

Signed-off-by: MichaIng <micha@dietpi.com>